Add KeyWindow class and refactor key display logic
