+++ .gitignore
@@ -1,0 +1,3 @@
+AnonChat/Client/file/session.txt
+AnonChat/Client/__pycache__/
+AnonChat/Client/file/
